refactor(android): unify duplicated AST segment splitting into shared utility
