+++ android/src/main/java/com/swmansion/enriched/markdown/MeasurementStore.kt
@@ -21,9 +21,11 @@ import com.swmansion.enriched.markdown.spans.MathMetrics
 import com.swmansion.enriched.markdown.spans.MathRenderMode
 import com.swmansion.enriched.markdown.styles.StyleConfig
 import com.swmansion.enriched.markdown.utils.common.FeatureFlags
+import com.swmansion.enriched.markdown.utils.common.MarkdownSegment
 import com.swmansion.enriched.markdown.utils.common.getBooleanOrDefault
 import com.swmansion.enriched.markdown.utils.common.getMapOrNull
 import com.swmansion.enriched.markdown.utils.common.getStringOrDefault
+import com.swmansion.enriched.markdown.utils.common.splitASTIntoSegments
 import com.swmansion.enriched.markdown.utils.text.extensions.replaceMathSpansWithPlaceholders
 import com.swmansion.enriched.markdown.views.TableContainerView
 import java.util.concurrent.ConcurrentHashMap
@@ -275,22 +277,6 @@ object MeasurementStore {
     return adjustedSize
   }
 
-/** Sealed interface for type-safe segment handling */
-  private sealed interface MarkdownSegment {
-    data class Text(
-      val nodes: List<MarkdownASTNode>,
-    ) : MarkdownSegment
-
-    data class Table(
-      val node: MarkdownASTNode,
-    ) : MarkdownSegment
-
-    data class Math(
-      val latex: String,
-      val node: MarkdownASTNode,
-    ) : MarkdownSegment
-  }
-
   private fun measureAndCacheSplit(
     context: Context,
     id: Int?,
@@ -425,44 +411,6 @@ object MeasurementStore {
           setUseLineSpacingFromFallbacks(true)
         }
       }.build()
-  }
-
-  private fun splitASTIntoSegments(root: MarkdownASTNode): List<MarkdownSegment> {
-    val segments = mutableListOf<MarkdownSegment>()
-    val currentTextNodes = mutableListOf<MarkdownASTNode>()
-
-    fun flushTextNodes() {
-      if (currentTextNodes.isNotEmpty()) {
-        segments.add(MarkdownSegment.Text(currentTextNodes.toList()))
-        currentTextNodes.clear()
-      }
-    }
-
-    for (child in root.children) {
-      when (child.type) {
-        MarkdownASTNode.NodeType.Table -> {
-          flushTextNodes()
-          segments.add(MarkdownSegment.Table(child))
-        }
-
-        MarkdownASTNode.NodeType.LatexMathDisplay -> {
-          flushTextNodes()
-          val latex =
-            if (child.children.isNotEmpty()) {
-              child.children.first().content
-            } else {
-              child.content
-            }
-          segments.add(MarkdownSegment.Math(latex, child))
-        }
-
-        else -> {
-          currentTextNodes.add(child)
-        }
-      }
-    }
-    flushTextNodes()
-    return segments
   }
 
   private fun tryRenderMarkdown(
